requirements.txt
numpy
pillow
tensorflow



APScheduler
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
requests

bing-image-downloader
scipy





gunicorn
whitenoise
dj-database-url
python-decouple

django-cloudinary-storage


